% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build Package}
\usage{
build(name)
}
\arguments{
\item{name}{Name of the package to build.}
}
\description{
This function documents and installs the specified package.
}
\examples{
build("ropufu")
}
\keyword{build,}
\keyword{document,}
\keyword{install,}
\keyword{reinstall}
