% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expspace.R
\name{expspace}
\alias{expspace}
\title{Exponentially Spaced Vector}
\usage{
expspace(from, to, n = 100)
}
\arguments{
\item{from}{The first element of the vector.}

\item{to}{The last element of the vector.}

\item{n}{Number of elements in the vector. Defaults to 100.}
}
\description{
This function creates a vector of exponentially spaced elements.
}
\examples{
expspace(1, 10, 8)
}
\keyword{linspace,}
\keyword{matlab,}
\keyword{seq}
