% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fills Region of a Plot}
\usage{
fill(x1, y1, x2, y2, ...)
}
\arguments{
\item{x1}{Vector of x-coordinates for the first curve.}

\item{y1}{Vector of y-coordinates for the first curve.}

\item{x2}{Vector of x-coordinates for the second curve.}

\item{y2}{Vector of y-coordinates for the second curve.}
}
\description{
This function fills a region of the plot.
}
\examples{
plot(1:10, 1:10)
fill(1:10, 1 + (1:10)^0.8, 1:10, sqrt(1:10), col = "gray", border = NA)
}
\keyword{fill,}
\keyword{plot,}
\keyword{polygon}
