% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linspace.R
\name{linspace}
\alias{linspace}
\title{Linearly Spaced Vector}
\usage{
linspace(from, to, n = 100)
}
\arguments{
\item{from}{The first element of the vector.}

\item{to}{The last element of the vector.}

\item{n}{Number of elements in the vector. Defaults to 100.}
}
\description{
This function creates a vector of linearly spaced elements.
}
\examples{
linspace(1, 10, 8)
}
\keyword{linspace,}
\keyword{matlab,}
\keyword{seq}
