% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchoosek.R
\name{nchoosek}
\alias{nchoosek}
\title{Binomial Coefficient}
\usage{
nchoosek(n, k)
}
\description{
This function calculates the binomial coefficient.
}
\examples{
nchoosek(10, 4)
}
\keyword{binomial,}
\keyword{choose,}
\keyword{combinatorial}
