% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.glm.R
\name{sas.glm}
\alias{sas.glm}
\title{Simple Linear Regression Plot}
\usage{
sas.glm(x, y, cl = 0.95, m = 100, fill.col = rgb(0, 0, 0, 0.1), ...)
}
\arguments{
\item{x}{The predictor variable.}

\item{y}{The response variable.}
}
\description{
This function creates a simple linear regression plot with prediction bounds.
}
\examples{
sas.glm(sort(runif(11, min = 0, max = 10)), 0:10 + rnorm(11), fill.col = "gray", frame = "F")
}
\keyword{GLM,}
\keyword{SAS,}
\keyword{plot}
\keyword{regression,}
